backend/requirements.txt
yt-dlp
# 用于HuggingFace风格分词器（AutoTokenizer）
modelscope
# 用于OpenAI tiktoken分词
tiktoken
# pytorch安装请去https://pytorch.org/get-started/locally/
# Windows：pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu126

fastapi
uvicorn
PyYAML
psycopg2-binary
python-dotenv

# 异步任务队列
celery
redis

pillow
# 用于重试机制，处理超时和限速
tenacity
aiohttp
requests
openai
litellm

# 向量数据库
chromadb

# ReAct 相关依赖
fastmcp

black
isort
pylint
pydantic_settings
python-multipart